+++ packages/core/src/configuration-non-customizable.ts
@@ -153,6 +153,7 @@ export const meter = {
 	},
 	total: {
 		paddingLeft: 36,
+		paddingRight: 24,
 	},
 	height: {
 		default: 8,